# Remove index persistency