--- a/index_test.go
+++ b/index_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"fmt"
 	"os"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -14,65 +13,77 @@ import (
 func TestIndex(t *testing.T) {
 	assert := require.New(t)
 
-	words := "The quick brown fox jumps over the lazy dog"
+	var idxtests = []struct {
+		in   string
+		addr Addr
+		ok   bool
+	}{
+		{"the", Addr{pos: 0, size: 3}, true},
+		{"quick", Addr{pos: 3, size: 5}, true},
+		{"brown", Addr{pos: 8, size: 5}, true},
+		{"fox", Addr{pos: 13, size: 3}, true},
+		{"jumps", Addr{pos: 16, size: 5}, true},
+		{"over", Addr{pos: 21, size: 4}, true},
+		{"missing", Addr{pos: 0, size: 0}, false},
+		{"the", Addr{pos: 0, size: 3}, true},
+		{"lazy", Addr{pos: 25, size: 4}, true},
+		{"dog", Addr{pos: 29, size: 3}, true},
+	}
+
 	var index []byte
 
-	t.Run("write", func(t *testing.T) {
-		idxRW := bytes.NewBuffer([]byte{})
-		idx, err := NewIndex(idxRW)
+	t.Run("open empty", func(t *testing.T) {
+		rw := bytes.NewBuffer([]byte{})
+		idx, err := NewIndex(rw)
 		assert.NoError(err)
+		assert.Len(idx.cache, 0)
 		assert.Equal(0, idx.cur)
-		pos := 0
-		for _, word := range strings.Fields(words) {
-			data := []byte(word)
+	})
+
+	t.Run("write", func(t *testing.T) {
+		rw := bytes.NewBuffer([]byte{})
+		idx, err := NewIndex(rw)
+		assert.NoError(err)
+		for _, tt := range idxtests {
+			if !tt.ok {
+				continue
+			}
+			data := []byte(tt.in)
 			size := len(data)
 			score := MakeScore(data)
-			expAddr := Addr{pos, size}
 			err := idx.Write(score, size)
 			assert.NoError(err)
-			assert.Equal(expAddr, idx.cache[score])
-			pos += size
+			before := idx.Len()
 			err = idx.Write(score, 0)
 			assert.NoError(err)
-			assert.Equal(expAddr, idx.cache[score], "Should ignore update")
+			assert.Equal(before, idx.Len(), "Should ignore same update")
 		}
-		index = make([]byte, idxRW.Len())
-		copy(index, idxRW.Bytes())
+		index = make([]byte, rw.Len())
+		copy(index, rw.Bytes())
 	})
 
 	t.Run("open", func(t *testing.T) {
-		tmp := make([]byte, len(index))
-		copy(tmp, index)
-		idxRW := bytes.NewBuffer(tmp)
-		idx, err := NewIndex(idxRW)
+		rw := bytes.NewBuffer(index)
+		idx, err := NewIndex(rw)
 		assert.NoError(err)
-		assert.Len(idx.cache, len(strings.Fields(words)))
-		assert.Equal(35, idx.cur)
+		assert.Len(idx.cache, 8)
+		assert.Equal(32, idx.cur)
 	})
 
 	t.Run("read", func(t *testing.T) {
-		tmp := make([]byte, len(index))
-		copy(tmp, index)
-		idxRW := bytes.NewBuffer(tmp)
-		idx, err := NewIndex(idxRW)
+		rw := bytes.NewBuffer(index)
+		idx, err := NewIndex(rw)
 		assert.NoError(err)
-		assert.Equal(35, idx.cur)
-		pos := 0
-		for _, word := range strings.Fields(words) {
-			data := []byte(word)
-			size := len(data)
-			score := MakeScore(data)
-			a, ok := idx.Read(score)
-			assert.Equal(pos, a.pos, "Should return correct position")
-			assert.Equal(size, a.size, "Should return correct size")
-			assert.True(ok, "Should indicate that score exists")
-			pos += size
+		for _, tt := range idxtests {
+			score := MakeScore([]byte(tt.in))
+			addr, ok := idx.Read(score)
+			if tt.ok {
+				assert.True(ok, "Should indicate that score exists")
+			} else {
+				assert.False(ok, "Should indicate that score is missing")
+			}
+			assert.Equal(tt.addr, addr, "Should return correct address")
 		}
-		score := MakeScore([]byte("missing"))
-		a, ok := idx.Read(score)
-		assert.Empty(a.pos, "Should return 0 position for missing score")
-		assert.Empty(a.size, "Should return 0 size for missing score")
-		assert.False(ok, "Should indicate that score doesn't exists")
 	})
 }