Convert tests into table driven and remove index persistency, coz +10% ns/op on open worth -40% decrease on write

```
$ benchcmp before.bench after.bench 
benchmark                   old ns/op      new ns/op     delta
BenchmarkOpenIndex-4        59921526       65754714      +9.73%
BenchmarkRebuildIndex-4     1151763798     64047897      -94.44%
BenchmarkOpen-4             59038646       64539848      +9.32%
BenchmarkWrite-4            20064          11718         -41.60%
BenchmarkRead-4             1696           1696          +0.00%

benchmark                   old allocs     new allocs     delta
BenchmarkOpenIndex-4        9672           9681           +0.09%
BenchmarkRebuildIndex-4     245483         9682           -96.06%
BenchmarkOpen-4             9676           9687           +0.11%
BenchmarkWrite-4            2              1              -50.00%
BenchmarkRead-4             1              1              +0.00%

benchmark                   old bytes     new bytes     delta
BenchmarkOpenIndex-4        31398202      31401925      +0.01%
BenchmarkRebuildIndex-4     35164512      31402700      -10.70%
BenchmarkOpen-4             31400805      31400878      +0.00%
BenchmarkWrite-4            1250          1234          -1.28%
BenchmarkRead-4             32            32            +0.00%
```